refactor plist to plaintext formatting (parse)
Plist parsing and formatting used by the old architecture is not
the way it should be handled now. It was not flexible and it was hard
to improve it.
It should be easier to test and setup the output formatting after the
changes.

Skip lists were not supported by the parse command, this
commit will add the skip files support for parse too.
