+++ libcodechecker/analyze/plist_parser.py
@@ -30,14 +30,17 @@ With the newer clang releases more information is available in the plist files.
 
 """
 
+from collections import defaultdict
+import math
 import os
 import plistlib
 import sys
 import traceback
 from xml.parsers.expat import ExpatError
 
+from libcodechecker import suppress_handler
+from libcodechecker import util
 from libcodechecker.logger import get_logger
-
 from libcodechecker.report import Report
 from libcodechecker.report import generate_report_hash
 
@@ -268,3 +271,183 @@ def skip_report_from_plist(plist_file, skip_handler):
         plist.seek(0)
         plist.write(new_plist_content)
         plist.truncate()
+
+
+class PlistToPlaintextFormatter(object):
+    """
+    Parse and format plist reports to a more human readable format.
+    """
+
+    def __init__(self,
+                 suppress_handler,
+                 skip_handler,
+                 severity_map,
+                 analyzer_type="clangsa"):
+
+        self.__analyzer_type = analyzer_type
+        self.__severity_map = severity_map
+        self.__print_steps = False
+        self.suppress_handler = suppress_handler
+        self.skiplist_handler = skip_handler
+
+    @property
+    def print_steps(self):
+        """
+        Print the multiple steps for a bug if there is any.
+        """
+        return self.__print_steps
+
+    @print_steps.setter
+    def print_steps(self, value):
+        """
+        Print the multiple steps for a bug if there is any.
+        """
+        self.__print_steps = value
+
+    @staticmethod
+    def __format_location(event, source_file):
+        loc = event['location']
+        line = util.get_line(source_file, loc['line'])
+        if line == '':
+            return line
+
+        marker_line = line[0:(loc['col'] - 1)]
+        marker_line = ' ' * (len(marker_line) + marker_line.count('\t'))
+        return '%s%s^' % (line.replace('\t', '  '), marker_line)
+
+    @staticmethod
+    def __format_bug_event(name, severity, event, source_file):
+
+        loc = event['location']
+        fname = os.path.basename(source_file)
+        if name:
+            return '[%s] %s:%d:%d: %s [%s]' % (severity,
+                                               fname,
+                                               loc['line'],
+                                               loc['col'],
+                                               event['message'],
+                                               name)
+        else:
+            return '%s:%d:%d: %s' % (fname,
+                                     loc['line'],
+                                     loc['col'],
+                                     event['message'])
+
+    def __write_bugs(self, output, reports, files, analyzed_source_file,
+                     report_stats):
+        """
+        Write out the bugs to the output and collect report statistics.
+        """
+
+        severity_stats = defaultdict(int)
+        file_stats = defaultdict(int)
+        report_count = defaultdict(int)
+
+        report_num = len(reports)
+        if report_num > 0:
+            index_format = '    %%%dd, ' % \
+                           int(math.floor(math.log10(report_num)) + 1)
+
+        non_suppressed = 0
+        for report in reports:
+            events = [i for i in report.bug_path if i.get('kind') == 'event']
+            f_path = files[events[-1]['location']['file']]
+            if self.skiplist_handler and \
+                    self.skiplist_handler.should_skip(f_path):
+                LOG.debug(report + ' is skipped (in ' + f_path + ")")
+                continue
+            hash_value = report.main['issue_hash_content_of_line_in_context']
+            bug = {'hash_value': hash_value,
+                   'file_path': f_path
+                   }
+            if self.suppress_handler and \
+                    self.suppress_handler.get_suppressed(bug):
+                LOG.debug("Suppressed by suppress file: {0}".format(report))
+                continue
+
+            last_report_event = report.bug_path[-1]
+            source_file = files[last_report_event['location']['file']]
+            report_line = last_report_event['location']['line']
+            report_hash = report.main['issue_hash_content_of_line_in_context']
+            checker_name = report.main['check_name']
+            sp_handler = suppress_handler.SourceSuppressHandler(source_file,
+                                                                report_line,
+                                                                report_hash,
+                                                                checker_name)
+
+            # Check for suppress comment.
+            suppress_data = sp_handler.get_suppressed()
+            if suppress_data:
+                if self.suppress_handler:
+                    hash_value, file_name, comment = suppress_data
+                    self.suppress_handler.store_suppress_bug_id(hash_value,
+                                                                file_name,
+                                                                comment)
+
+                continue
+
+            file_stats[f_path] += 1
+            severity = self.__severity_map.get(checker_name,
+                                               'UNSPECIFIED')
+            severity_stats[severity] += 1
+            report_count["report_count"] += 1
+
+            output.write(self.__format_bug_event(checker_name,
+                                                 severity,
+                                                 last_report_event,
+                                                 source_file))
+            output.write('\n')
+            output.write(self.__format_location(last_report_event,
+                                                source_file))
+            output.write('\n')
+
+            if self.print_steps:
+                output.write('  Report hash: ' + report_hash + '\n')
+                output.write('  Steps:\n')
+                for index, event in enumerate(events):
+                    output.write(index_format % (index + 1))
+                    source_file = files[event['location']['file']]
+                    output.write(self.__format_bug_event(None,
+                                                         None,
+                                                         event,
+                                                         source_file))
+                    output.write('\n')
+            output.write('\n')
+
+            non_suppressed += 1
+
+        basefile_print = (' ' +
+                          os.path.basename(analyzed_source_file)) \
+            if analyzed_source_file and \
+            len(analyzed_source_file) > 0 else ''
+
+        if non_suppressed == 0:
+            output.write('Found no defects while analyzing%s\n' %
+                         (basefile_print))
+        else:
+            output.write(
+                'Found %d defect(s) while analyzing%s\n\n' %
+                (non_suppressed, basefile_print))
+
+        report_stats["severity"] = severity_stats
+        report_stats["files"] = file_stats
+        report_stats["reports"] = report_count
+
+    def parse_and_write(self, plist_file, analyzed_source_file,
+                        output=sys.stdout):
+        """
+        Parse a plist report file format it to a more human readable format.
+        """
+        report_stats = {}
+        try:
+            files, reports = parse_plist(plist_file)
+        except Exception as ex:
+            traceback.print_stack()
+            LOG.error('The generated plist is not valid!')
+            LOG.error(ex)
+            return 1
+
+        self.__write_bugs(output, reports, files, analyzed_source_file,
+                          report_stats)
+
+        return report_stats
